!24. рефакторинг 1. добавил пустые строки в конец файла где их не было. 2. перенес проверки из Dao в Ресурс 3. заменил !isBlank на isNotBlank
