+++ nginx/conf.d/app.conf
@@ -1,4 +1,4 @@
-    server {
+server {
      listen 80;
      charset utf-8;
      access_log off;
@@ -15,6 +15,4 @@
        proxy_set_header  X-Forwarded-For $proxy_add_x_forwarded_for;
        proxy_set_header  Host $http_host;
      }
-
-
 }
