+++ nginx/www/html/style.css
@@ -1,4 +1,4 @@
-﻿div {
+div {
   margin: 8px;
   padding-left: 8px;
   padding-right: 20px;
