+++ src/main/java/ru/hh/todolist/dao/TaskDao.java
@@ -1,25 +1,18 @@
 package ru.hh.todolist.dao;
 
 import jakarta.inject.Inject;
-import jakarta.ws.rs.BadRequestException;
 import org.hibernate.SessionFactory;
-import org.slf4j.Logger;
 import org.springframework.stereotype.Repository;
 import ru.hh.todolist.entity.Task;
 import ru.hh.todolist.entity.TaskStatus;
-import ru.hh.todolist.service.TodoListService;
 import ru.hh.todolist.utils.TransactionHelper;
 
 import java.util.List;
-
-import static org.slf4j.LoggerFactory.getLogger;
 
 @Repository
 public class TaskDao {
   private final SessionFactory sessionFactory;
   private final TransactionHelper transactionHelper;
-  public static final Logger LOGGER = getLogger(TodoListService.class);
-
 
   @Inject
   public TaskDao(SessionFactory sessionFactory, TransactionHelper transactionHelper) {
@@ -37,14 +30,12 @@ public class TaskDao {
   }
 
   public Task get(Long taskId) {
-    checkId(taskId);
     return transactionHelper.inTransaction(() ->
         sessionFactory.getCurrentSession().get(Task.class, taskId)
     );
   }
 
   public void delete(final Long taskId) {
-    checkId(taskId);
     transactionHelper.inTransaction(() ->
         sessionFactory.getCurrentSession().createQuery(
                 "DELETE from Task as t WHERE t.id = :taskId")
@@ -68,12 +59,5 @@ public class TaskDao {
                 "SELECT t FROM Task as t ORDER BY t.id", Task.class)
             .list()
     );
-  }
-
-  private static void checkId(Long taskId) {
-    if (taskId == null || taskId < 0) {
-      LOGGER.error("TaskId parameter is negative or Null: {}", taskId);
-      throw new BadRequestException("TaskId bad.");
-    }
   }
 }
